requirements.txt
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=0.24.0
mysql-connector-python>=8.0.0
python-dotenv>=0.19.0
mlxtend>=0.22.0
xgboost>=2.0.0
jupyter>=1.0.0
statsmodels>=0.13.0 